AI/ML/DL/PyTorchBasics/requirements.txt
# for completing assignment
filetype
numpy=1.26.4
pandas=2.2.3
pytorch=2.4.1

# for FRQs and test cases only
matplotlib=3.9.2
pytest
